+++ src/app/(providers)/facilities/page.module.scss
@@ -1,51 +1,51 @@
-.mapContainer {
-  position: relative;
-}
+// .mapContainer {
+//   position: relative;
+// }
 
-.currentLocation {
-  width: 45px;
-  height: 45px;
-  border-style: none;
-  border-radius: 10px;
-  background-color: rgb(255, 255, 255);
-  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.5);
-  z-index: 2;
-  position: absolute;
-  bottom: 5%;
-  left: 95%;
-  font-size: 20px;
-  line-height: 50px;
-  cursor: pointer;
-}
+// .currentLocation {
+//   width: 45px;
+//   height: 45px;
+//   border-style: none;
+//   border-radius: 10px;
+//   background-color: rgb(255, 255, 255);
+//   box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.5);
+//   z-index: 2;
+//   position: absolute;
+//   bottom: 5%;
+//   left: 95%;
+//   font-size: 20px;
+//   line-height: 50px;
+//   cursor: pointer;
+// }
 
-.currentLocation:hover {
-  background-color: #868686;
-  color: #ffffff;
-}
+// .currentLocation:hover {
+//   background-color: #868686;
+//   color: #ffffff;
+// }
 
-// 반응형
+// // 반응형
 
-@media screen and (max-width: 1200px) {
-  .mapContainer {
-    width: 100%;
-  }
-  .currentLocation {
-    left: 90%;
-  }
-}
+// @media screen and (max-width: 1200px) {
+//   .mapContainer {
+//     width: 100%;
+//   }
+//   .currentLocation {
+//     left: 90%;
+//   }
+// }
 
-@media screen and (max-width: 820px) {
-  .mapContainer {
-    width: 100%;
-  }
-  .currentLocation {
-    left: 90%;
-  }
-}
+// @media screen and (max-width: 820px) {
+//   .mapContainer {
+//     width: 100%;
+//   }
+//   .currentLocation {
+//     left: 90%;
+//   }
+// }
 
-@media screen and (max-width: 480px) {
-  .currentLocation {
-    left: 8%;
-    bottom: 90%;
-  }
-}
+// @media screen and (max-width: 480px) {
+//   .currentLocation {
+//     left: 8%;
+//     bottom: 90%;
+//   }
+// }
