+++ src/app/_components/facilities/facilitiesMapComponent.module.scss
@@ -3,6 +3,7 @@
   height: 90vh;
   z-index: 1;
   margin: 0;
+  position: relative;
 }
 
 .mapWrap {
@@ -23,49 +24,16 @@
   }
 }
 
-.overlayWrap {
-  position: absolute;
-  left: 0;
-  bottom: 40px;
-  width: 320px;
-  height: 120px;
-  margin-left: -154px;
-  text-align: left;
-  font-size: 12px;
-  line-height: 1.6;
-  background-color: #ffffff;
-  border-radius: 20px;
-  border: 1px solid #0ac4b9;
-  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.3);
-  padding: 10px 5px;
+// 반응형
+
+@media screen and (max-width: 1200px) {
+  .mapContainer {
+    width: 100%;
+  }
 }
 
-.placeName {
-  height: 30px;
-  border-bottom: 1px solid #ddd;
-  padding: 10px;
-  font-size: 17px;
-  font-weight: 500;
-  display: flex;
-  align-items: center;
-}
-
-.placeContent {
-  padding: 8px 10px;
-}
-
-.address {
-  font-weight: 700;
-}
-
-.link {
-  color: #0ac4b9;
-  line-height: 20px;
-  font-size: 12px;
-  line-height: 15px;
-}
-
-.currentMarker {
-  padding: 5px;
-  text-align: center;
+@media screen and (max-width: 820px) {
+  .mapContainer {
+    width: 100%;
+  }
 }
