+++ src/app/_components/facilities/facilitiesOverlay.module.scss
@@ -1,0 +1,30 @@
+.overlayWrap {
+  position: absolute;
+  left: 0;
+  bottom: 40px;
+  width: 320px;
+  height: 120px;
+  margin-left: -154px;
+  text-align: left;
+  font-size: 12px;
+  line-height: 1.6;
+  background-color: #ffffff;
+  border-radius: 20px;
+  border: 1px solid #0ac4b9;
+  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.3);
+  padding: 10px 5px;
+}
+
+.placeName {
+  height: 30px;
+  border-bottom: 1px solid #ddd;
+  padding: 10px;
+  font-size: 17px;
+  font-weight: 500;
+  display: flex;
+  align-items: center;
+}
+
+.placeContent {
+  padding: 8px 10px;
+}
